ai-model-hub/backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
litellm==0.6.0
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
redis==4.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
openai==1.0.0
tiktoken==0.4.0
asyncio==3.4.3
bcrypt==4.0.1
PyJWT==2.8.0
aiosqlite==0.19.0 # For development DB if needed

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting & Formatting
black==23.3.0
isort==5.12.0
flake8==6.0.0 